refactor: rename the module design_creator to design_creator_module
