+++ elasticai/creator/nn/fixed_point/hard_sigmoid/layer.py
@@ -1,5 +1,5 @@
 from elasticai.creator.base_modules.hard_sigmoid import HardSigmoid as HardSigmoidBase
-from elasticai.creator.nn.design_creator import DesignCreatorModule
+from elasticai.creator.nn.design_creator_module import DesignCreatorModule
 from elasticai.creator.nn.fixed_point._two_complement_fixed_point_config import (
     FixedPointConfig,
 )
