+++ elasticai/creator/nn/fixed_point/precomputed/precomputed_module.py
@@ -2,7 +2,7 @@ from typing import cast
 
 import torch
 
-from elasticai.creator.nn.design_creator import DesignCreatorModule
+from elasticai.creator.nn.design_creator_module import DesignCreatorModule
 from elasticai.creator.nn.fixed_point._math_operations import MathOperations
 from elasticai.creator.nn.fixed_point._two_complement_fixed_point_config import (
     FixedPointConfig,
